backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
requests==2.31.0
prophet==1.1.5
cmdstanpy==1.2.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

